requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
docker==7.1.0
pyyaml==6.0.3
jinja2>=3.1.6
python-multipart>=0.0.20